Fix: 결제 파트 프론트 에러 해결 (#214)
* fix: int -> long

* fix: oauth2 도메인 설정

* refactor: 로그 한 줄 추가

* fix: 도메인 문제....

* test: 테스트 제거

* refactor: AOP Proceed 제거

+++ src/main/java/com/jelly/zzirit/domain/order/dto/request/PaymentRequest.java
@@ -21,7 +21,7 @@ public record PaymentRequest(
 
 	@Schema(description = "총 결제 금액 (단위: 원)", example = "15000")
 	@Positive(message = "결제 금액은 0보다 커야 합니다.")
-	int totalAmount,
+	long totalAmount,
 
 	@Schema(description = "배송 요청사항", example = "문 앞에 놔주세요")
 	String shippingRequest,
